requirements.txt
appdirs==1.4.4
astroid==2.4.2
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
Click==7.0
colorama==0.4.3
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
idna==2.10
isort==5.5.0
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.6.0
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.12
toml==0.10.1
urllib3==1.25.11
virtualenv==20.0.35
Werkzeug==0.16.0
wrapt==1.12.1
dateparser==1.0.0
python-dateutil>=2.6.0
selenium>=3.4.3
python-axolotl>=0.2.3
cryptography>=2.9.2
typing>=3.7.4.1
numpy>=1.18.4
python-magic>=0.4.18
python-resize-image==1.1.19
Pillow>=7.1.2
timg==1.1.4
scikit-image==0.17.2
imgur-uploader==0.3.0
spotipy==2.16.1
youtube-search==1.1.1
lxml==4.6.2
google==3.0.0
google_trans_new==1.1.9
youtube-title-parse==1.0.0
yfinance==0.1.55
cryptocmd==0.6.0
emoji==0.6.0
ShazamAPI==0.0.2
pydub==0.24.1
SpeechRecognition==3.8.1
opencv-python==4.2.0.32
websocket_client==0.57.0
pyopenssl==20.0.1
pyttsx3==2.90
gTTS==2.2.2
playsound==1.2.2
flask-socketio==5.0.3
munch==2.5.0
